requirements.txt
python-dotenv
pandas
secure-smtplib
gspread
google-auth